ktrex_stateHandlerA10: u8 cond branchy bool
94.1->96.6%. Same lever as A02: the state-equality bool 'cond' (unkFF==K
chain) is u8 so the && test re-masks via clrlwi+cmplwi to match target
(#38/#58). Residual: single-use lfsx re-fold (#112) and push-store
positioning.

(cherry picked from commit a6854c1de3aa47e6347759c3fdd5192b05f140d1)

+++ src/main/dll/DR/dll_0250_ktrex.c
@@ -1703,7 +1703,7 @@ int ktrex_stateHandlerA10(int obj, int runtime)
     {
         if ((((KTRexArenaState*)gKTRexState)->timerFA & 8) != 0)
         {
-            int cond;
+            u8 cond;
             u8 fe;
             ((KTRexArenaState*)gKTRexState)->unk101 += 1;
             GameBit_Set(0x572, ((KTRexArenaState*)gKTRexState)->unk101);
